requirements.txt
asgiref==3.7.1
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-js-asset==2.0.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
numpy==1.24.3
outcome==1.2.0
pandas==2.0.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
wsproto==1.2.0
